create_agent (1).py
# Import necessary libraries
import autogpt
import tensorflow as tf
from tensorflow import keras
import numpy as np
import pandas as pd
import requests
from bs4 import BeautifulSoup
import nltk
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.model_selection import train_test_split
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import accuracy_score
import matplotlib.pyplot as plt
import seaborn as sns
import gensim
from gensim.models import Word2Vec
import pydot
import graphviz
from IPython.display import Image
from IPython.display import display
from IPython.display import HTML
from tqdm import tqdm
from PIL import Image
import pytesseract
from pdfminer.high_level import extract_text
from docx import Document
from pptx import Presentation
import win32com.client
import os
import sys
import shutil
import subprocess
import time
import random

# Create the AI agent
agent = autogpt.Agent('python_agent', 'AI agent that writes Python code, including downloading dependencies and running the code.')