requirements (1).txt
autogpt
tensorflow
keras
numpy
pandas
beautifulsoup4
requests
nltk
scikit-learn
matplotlib
seaborn
gensim
pydot
graphviz
ipython
jupyter
ipykernel
ipywidgets
tqdm
pillow
opencv-python
pytesseract
pdfminer.six
python-docx
python-pptx
pywin32
pyinstaller
